Add support for retry of connection to the database.
At some point we might want to retry connecting the connection to
the database server in case we had some transient error, or if
connections on a standby, or even a primary, are not still available
because the server is in recovery and hadn't found a consistent state
yet.

This isn't the most elegant way of dealing with this issue, but it
looks like the best option with the what we have available.

In this commit we only add relevant artifacts we'll use later in the
repmgr-action-standby.c switchover code, which has been giving
some trouble already.